requirements.txt
Flask==2.2.5
SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.2.3
reportlab==4.0.8
arabic-reshaper==3.0.0
python-bidi==0.4.2
gunicorn==21.2.0
